GuestPropertySvc.h,Main: Changed GuestPropValidateName and GuestPropValidateValue to take size_t instead of uint32_t, eliminating the need for a bunch of (uint32_t) casts. Call setErrorBoth when either of the two fails in Main.  bugref:10185
svn:sync-xref-src-repo-rev: r151930

+++ src/VBox/Main/xml/Settings.cpp
@@ -1,4 +1,4 @@
-/* $Id: Settings.cpp 95337 2022-06-21 20:50:20Z knut.osmundsen@oracle.com $ */
+/* $Id: Settings.cpp 95338 2022-06-21 21:20:50Z knut.osmundsen@oracle.com $ */
 /** @file
  * Settings File Manipulation API.
  *
@@ -4656,7 +4656,7 @@ void MachineConfigFile::readGuestProperties(const xml::ElementNode &elmGuestProp
         /* Check guest property 'name' and 'value' for correctness before
          * placing it to local cache. */
 
-        int rc = GuestPropValidateName(prop.strName.c_str(), (uint32_t)prop.strName.length() + 1  /* '\0' */);
+        int rc = GuestPropValidateName(prop.strName.c_str(), prop.strName.length() + 1  /* '\0' */);
         if (RT_FAILURE(rc))
         {
             LogRel(("WARNING: Guest property with invalid name (%s) present in VM configuration file. Guest property will be dropped.\n",
@@ -4664,7 +4664,7 @@ void MachineConfigFile::readGuestProperties(const xml::ElementNode &elmGuestProp
             continue;
         }
 
-        rc = GuestPropValidateValue(prop.strValue.c_str(), (uint32_t)prop.strValue.length() + 1  /* '\0' */);
+        rc = GuestPropValidateValue(prop.strValue.c_str(), prop.strValue.length() + 1  /* '\0' */);
         if (rc == VERR_TOO_MUCH_DATA)
         {
             LogRel(("WARNING: Guest property '%s' present in VM configuration file and has too long value. Guest property value will be truncated.\n",
